client: add try_from_event_handler_and_connection_config of ClientConnectionConfig
Downstream implementing can easier to resue method of ClientConnectionConfig.
